+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-06-29 21:09:06
-// Commit: feat(state): add async state stream provider
+// Updated at: 2026-06-30 10:38:37
+// Commit: test(state): add unit tests for state business logic
 
 class stateComponentState {
-  final String commitRef = '2413';
-  final String lastMessage = 'feat(state): add async state stream provider';
-  final DateTime timestamp = DateTime.parse('2026-06-29T21:09:06+05:00');
+  final String commitRef = '2422';
+  final String lastMessage = 'test(state): add unit tests for state business logic';
+  final DateTime timestamp = DateTime.parse('2026-06-30T10:38:37+05:00');
 }